+++ CHANGELOG.rst
@@ -2,9 +2,10 @@ Changelog for package franka_ros2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 UNRELEASED
--------------------
+----------
 Requires libfranka >= 0.18.0 and franka_description >= 2.0.0 requires ROS 2 Jazzy
 
+* Add: Added a joint-based point-to-point motion action with usage example
 * BREAKING CHANGE: arm_id replaced by robot_type and controller_name by controller_names
 * Feat: TMRv0.2 teleoperation example controller added
 * Feat: arm_id replaced by robot_type